requirements.txt
# LangChain Core
langchain
langchain-core

# LangGraph Core
langgraph

# Google Gemini (PaLM) Integration
langchain-google-genai
google-generativeai

# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub

# Environment Variable Management
python-dotenv


# Machine Learning Utilities
numpy
scikit-learn

pandas
streamlit
numpy==1.26.4

# pipreqs . --force --ignore chatbotvenv